requirements.txt
# click>=8.1.8
# # Keep crewai at version 0.130.0
# # Other package versions are flexible and can be changed
# # Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
# crewai==0.130.0 
# crewai-tools==0.47.1
# fastapi==0.110.3
# google-ai-generativelanguage==0.6.4
# # google-api-core==2.10.0
# google-api-python-client==2.131.0
# google-auth==2.29.0
# google-auth-httplib2==0.2.0
# google-cloud-aiplatform==1.53.0
# google-cloud-bigquery==3.23.1
# google-cloud-core==2.4.1
# google-cloud-resource-manager==1.12.3
# google-cloud-storage==2.16.0
# google-crc32c==1.5.0
# google-generativeai==0.5.4
# google-resumable-media==2.7.0
# googleapis-common-protos==1.63.0
# Jinja2==3.1.4
# jsonschema==4.22.0
# langchain-core==0.1.52
# langsmith==0.1.67
# numpy==1.26.4
# oauthlib==3.2.2
# onnxruntime==1.22.0
# openai==1.30.5
# opentelemetry-api>=1.30.0
# opentelemetry-exporter-otlp-proto-common==1.25.0
# opentelemetry-exporter-otlp-proto-grpc==1.25.0
# opentelemetry-exporter-otlp-proto-http>=1.30.0
# opentelemetry-instrumentation==0.46b0
# opentelemetry-instrumentation-asgi==0.46b0
# opentelemetry-instrumentation-fastapi==0.46b0
# opentelemetry-proto==1.25.0
# opentelemetry-sdk>=1.30.0
# opentelemetry-semantic-conventions==0.46b0
# opentelemetry-util-http==0.46b0
# pandas==2.2.2
# pillow==10.3.0
# pip==24.0
# protobuf==4.25.3
# pydantic>=2.4.2
# pydantic_core==2.8.0


# click>=8.1.8

# # Core AI
# crewai==0.130.0
# crewai-tools==0.47.1

# # API
# fastapi==0.110.3

# # Google
# google-generativeai==0.5.4
# google-ai-generativelanguage==0.6.4
# google-api-python-client==2.131.0
# google-auth==2.29.0
# google-auth-httplib2==0.2.0

# # Data
# pandas==2.2.2
# numpy==1.26.4
# pillow==10.3.0

# # Validation
# pydantic>=2.4.2

# # Utils
# jinja2==3.1.4
# jsonschema==4.22.0
# oauthlib==3.2.2


# crewai==0.130.0
# crewai-tools==0.47.1
# fastapi
# uvicorn[standard]
# python-dotenv
# python-multipart
# langchain
# langchain-community
# pypdf

## Core CrewAI
crewai==0.130.0
crewai-tools==0.47.1

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

# LangChain (FIXED VERSION RANGE)
langchain>=0.3.7,<0.4.0
langchain-community>=0.3.7,<0.4.0

# PDF reader (required by embedchain)
pypdf>=5.0.0,<6.0.0

# Queue system
celery>=5.4.0
redis>=5.0.7

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# CrewAI dependency stabilization
litellm>=1.44.0
openai>=1.40.0
tiktoken>=0.8.0

# Required by CrewAI memory system
chromadb>=0.5.23

# Utilities
requests>=2.32.3